Changing the Merb::Rake::Task::BINARY constant for a configurable parameter, this way frameworks like Merb can configure the location of the binary file.
Also there was added another Rake tasks that allows people to easily invoke an specific feature using only the command (with default name feature):

$ rake feature[feature_name]

This is equivalent to using:

$ rake features FEATURES=feature_name

But the latter would be more comfortable visually and also would be more easy to invoke from a terminal.

This idea was taken from the merb_cucumber gem.