intelligent handling of profile VS command line args
Default formatters defined in profiles are now ignored if the command
line args specify one to override it.

The logic to handle all of the special cases of how to merge command
line args and profiles was getting very complex.  In order to keep the
code simple I refactored the CLI quite a bit.  I extracted two classes
from Configuration: Options and ProfileLoader.  The responsibilities are
much better handled by the new classes and quite a bit of cruft was
removed.

I have not removed/moved all the redundant specs from Configuration yet.  I
plan on doing that later but as of now the refactor is complete.