348-shared-examples: Shared Examples in Scenario Outlines
This patch allows multiple Scenario Outlines to share the same Examples.
The general consensus on issue #348 favored putting the Examples block
at the end of all Scenario Outlines:

  https://rspec.lighthouseapp.com/projects/16211-cucumber/tickets/348

Such is the way it was implemented; all Scenario Outlines use the
nearest Examples block that follows it. This was achieved through a
marginally janky approach of keeping track of any SO's with missing
examples, as the ScenarioOutline class attribute @@missing_examples,
then backfilling those SO's once an Examples block is reached. This
seemed the most straightforward way to do it, as it does not involve any
changes to the Gherkin syntax.

I've included tests in features/shared_examples.feature.