Support specifying step definitions by symbol.
If a symbol is passed to a step definition, then that symbol will be
sent to the World object. The symbol must therefore refer to a method
defined by the World: either in an included module or on the World
object.

The proc can still be used, but if a symbol is provided the proc will
be ignored.

Also, support executing the symbol against objects other than the World.

If an :on key is provided to the options hash then the proc attached
to this will be evaluated against the World, and then the symbol sent
to the result of that proc, rather than directly to the World itself.