Fall back on Test::Unit::Assertions if rspec-expectations is missing
Currently, users who don't want to use rspec-expectations have to do a
little extra setup in their env.rb to use Test::Unit.  Since Test::Unit
ships with Ruby (albeit as a wrapper in 1.9), it makes sense to look
for this library as a fallback.  This lowers the barrier even further
for users who want to jump straight into Cucumber with minimal
dependencies.