Refactor: remove apparently dead code
No tests broke when I removed this, and I was perplexed by what it might
have been intended to do, since it would have apparently set all
subsequent scenarios in a feature to be skipped if one of them failed.

I don't really understand why it didn't, but it's probably to do with
how step invocations are cached on steps, with the step invocations
being created anew when the next scenario is visited.