Add spike binary that integrates with new 2.0 core.
This commit represents a major milestone. We can now run
(some of) Cucumber's own tests using the new core, with this
spike front-end.

Try it!

```
./bin/cuke features/docs/getting_started.feature
```

We now need to work our way through Cucumber's features until
we can run all of them using this front-end, gradually adding new
features to Cucumber-Core as we need them.

Once all the tests are passing and we know the core at least supports
the features needed to run Cucumber's own tests (but still running them
against the current stable version of Cucumber), then we can start
trying to integrate it with the real front-end, and running the tests
against the new code itself.

@tooky @os97673 let's get to it!