Aha. Now printing steps of a scenario outline.
This represents learning for me. The `describe_source_to` method won't
-- currently anyway -- walk down into each node in the source. So when
you get a `ScenarioOutline` from `describe_source_to` you still need to ask
it to describe itself so you can see the `OutlineStep`s within it.

This will be the same for `Ast::Step` source nodes where we want to see
the Multiline arg within it.