junit formatter specs all passing.
@mattwynne @tooky

While getting the junit formatter spec working I have noticed somthing strange.

For a Examples table any repeated undefined rows are ignored and are not
displayed.

For example

```cucumber
Feature: A Feature
  Scenario Outline:
    Given there are <Things>

    Examples: Fruit
      | Things  |
      | apples  |
      | apples  |
      | bananas |
```

When `USE_LEGACY` is on, the formatter output will have a table like:

```cucumber
| Things  |
| apples  |
| apples  |
| bananas |
```

When `USE_LEGACY` is off, the formatter output will have a table like:

```cucumber
| Things  |
| apples  |
| bananas |
```

I have tested this with the junit and html formatters.

The pretty formatter displays

```cucumber
| Things  |
| apples  |
| bananas |
```

regardless of `USE_LEGACY`.