Decode base64 data before sending it to a gherkin formatter
The gherkin JSON formatter will encode the data it receives before
inserting it into the json file. When supporting embedding images in
html files, the data sent to the embed method need to be base64
encoded. Unless decoded before sent to the gherkin JSON formatter, the
base64 encoding will be applied twice.