Fix output when table diffs fail.
I don't use this feature myself, so I'm not familiar with it, but from
the look of this acceptance test, we were MUTATING THE AST OBJECT, THEN
RENDERING IT.

My head hurts.

I'm just printing the nice diff output in the exception now, which I
think reads better and also doesn't involve so much crazy coupling to
the AST / formatters.