Undefined nested steps
This scenario highlights a regression about how we print snippets for
undefined, nested steps.

What's the right behaviour here? Previously, we treated these as the
same as undefined steps from features, but now I've looked into this,
I'm not so sure.

This scenario describes the existing behaviour, but is it desirable?

If you have a nested step, surely you expect it to already match? If
it fails to match, I think you just want that to be a straight up
failure, rather than a regular "undefined" error that prints a snippet.

In the new 2.0 model, we know about undefined Gherkin steps as soon as we've
compiled the mappings, but we can't know about nested steps until we
execute the action, so they're different. It feels like the model might
be telling us something.