Reset previous_step_keyword at the start of each test case
Let the Pretty and Progress formatter collect the actual keywords and
steps in the Console module, instead of asking the runtime for the step
in the Console#print_snippets. Use before_test_case to reset the
previous_step_keyword at the start of each test case. Remove the
language from the StepInvocation, it is not needed when using
Cucumber::Core::Ast::Step#actual_keyword