Prepare world as an around hook.
This ensures that prepare world is the first thing that is
run for a test case, preventing the inconsistency with what the world _is_
in around hooks.

Legacy adapter requires each test case to have a test step otherwise it
doesn't switch step container and messages get lost.

Also ensures failures in Around hooks are reported.