Ensure each test case has a test step.
Legacy adapter requires each test case to have a test step otherwise it
doesn't switch step container and messages get lost.

@mattwynne I think we might be able to deal with this if we switch step
container in before_test_case, but it's something I would prefer to pair
on you with? Seems like a big change, and maybe not something we need to
worry about for 2.0?