Correct the calculation of hook locations. Fixes #885.
It is not possible to use relative_path_from on Windows the two paths
are located on different drives ('c:', 'd:' etc.). For hooks base the
location on the file_colon_line method created for the Proc in
cucumber/core_ext/proc.rb. For step definitions with symbols, calculate
the path the same way as in cucumber/core_ext/proc.rb. Also fix that
when slitting up a 'file_colon_line', the file on Windows may contain
a ':' (c:\path\file.ext), so split(':') does not work properly in all
cases.