Moved simplecov_setup into lib folder
As simplecov_setup is a runtime dependency of the cucumber bin script it
should not be placed into spec/ folder as there should be only test
parts be placed. The old code bites us at SUSE as all tests are split
into a testsuite package which results in a currently broken cucumber.