Use a random seed that works on all versions
The implementation of Array#shuffle in Ruby 2.x differs from that used
in JRuby and Ruby 1.9.3. This seed happens to produce the same results
on both.

This is fragile in that it might break in future versions of Ruby. A
better solution would be to change the shuffle implementation to one
that is not implementation-specific.