Make seeded random order platform-independent
With this change, using the --order=random option with a particular seed
will produce the same order on every Ruby interpreter.

Previously, the random seed would produce consistent results on a given
interpreter, but this could not always be reproduced on another version
of Ruby. This occurred because Array#shuffle is implemented differently
in different interpreters: Ruby 1.9.3 and JRuby 1.7.x share one
implementation, whilst Ruby 2.x has a different one.