Run scenarios in fully random order (#970)
This documents the desired behaviour of the random order option across
multiple features and corrects the ordering of filters within runtime.rb
in order to run scenarios in a genuinely random order.

Previously, using --order=random resulted in the order of scenarios
being randomised within a feature, but features themselves always
proceeded in alphabetical (or ASCIIbetical) order based on their file
path, i.e.:

    Feature: 1
      Scenario: 1B
      Scenario: 1A

    Feature: 2
      Scenario: 2B
      Scenario: 2A

Although Randomizer does indeed shuffle the scenarios, it was followed
in the filter chain by LocationsFilter which groups scenarios by file
path. Moving Randomizer later in the chain avoids this.

Now, each scenario is run in random order regardless of the name or path
of the feature:

    Feature: 1
      Scenario: 1A

    Feature: 2
      Scenario: 2B
      Scenario: 2A

    Feature: 1
      Scenario: 1B

See also https://github.com/cucumber/cucumber-ruby/issues/969

The implementation of Array#shuffle in Ruby 2.x differs from that used
in JRuby and Ruby 1.9.3. This seed happens to produce the same results
on both.

This is fragile in that it might break in future versions of Ruby. A
better solution would be to change the shuffle implementation to one
that is not implementation-specific.