Specify that random order is stable
In addition to predictability, another useful property of random
ordering is stability, i.e. if a given seed results in the order BCDA,
then skipping a step will not affect the relative order of the other
steps: if C is skipped, the order should be BDA. This is a necessary
precondition for bisection, should we wish to add that in the future.

As randomisation takes place before filtering we get stability for free.
This change just documents it and preserves it from future regression.