First pass at spec formatter
This formatter provides a high-level summary of the scenarios that were
run. It's also A LOT simpler than the pretty formatter, and so more
reliable for testing with.

I'm thinking we can use this to build out alternative re-usable pieces
(like the summary) that need to be re-written anyway.

Work still to do:

- unit tests
- acceptance test for scenario outlines
- finish the summary
- colours?
- some animation as steps run?