Fix HTML injection in HTML formatter message output
HTML formatter appends any scenario output to the HTML without
escaping HTML tags in the messages. This means that any output
from the steps gets injected into the report page.

In the simple case, this won't display the message as expected.
When combined with for example capybara and CI environments this
opens up the possibility of an XSS attack from the user input on
the environment being tested.

Using manual escaping of the output itself would solve the HTML
injection, but since the same output is used by multiple formatters
it no longer prints the actual message in non-HTML formatters.