Use single quotes in generated template steps
When pending steps are generated for undefined steps, the step name does
not include any interpolation so double-quotes are not necessary. Only
the single-quotes are necessary for those strings.

This will make the generated steps allow consistent with the ruby style
guide:
  https://github.com/rubocop-hq/ruby-style-guide#strings

And with the default settings for Rubocop Style/StringLiterals cop:
  https://github.com/rubocop-hq/rubocop/blob/ab6346ed37d88a4b900c96979f8ef18959e8dcce/lib/rubocop/cop/style/string_literals.rb