Fixes env.rb and support files load order
According to The Cucumber Book on pp. 129-130:

> Just like `features/step_definitions`, Cucumber will load all Ruby
> files it finds in `features/support`. This is convenient, because it
> means you don't need to pepper `require` statements everywhere, but it
> does mean that the precise order in which files are loaded is hard to
> control. In practice, this means you can't have dependencies between
> files in the `support` directory, because you can't expect that
> another file will have already been loaded. There is one exception to
> that, a special file called `env.rb`.
>
> The file `features/support/env.rb` is always the very first file to be
> loaded when Cucumber starts a test run.

This is also extened such that any `env.rb` file located in descendant
`support` folders are _also_ loaded before other `*.rb` files in
`support` folders.

However, when attempting to create a better port of Cucumber in the
Pester PowerShell testing framework, I found this load ordering to not
be correct.

This commit fixes the order in which files in `support` folders are
loaded so that:

1. the `features/support/env.rb` file is always loaded first, if found
2. other `env.rb` files located in descendent `support` folders under
`features` are loaded second
3. all other `*.rb` files in descendant `support` folders are loaded