Updates require option help text
When I read the help text for the --require option, it was conceivable
that the text could have been interpreted in at least two ways.

This commit aims to clarify the help text to reduce ambiguity/improve
the clarity of the behavior of the option, espcially with respect to the
loading of files under "support" directories.