CI/Win - try another approach to clean-up tmp/aruba
We don't have any error displayed, but previous run shown that, some directories
are kept. My guess is that they are not removed as they are not empty. I was not able
to reproduce that, but I got errors on Windows saying the directory can not be removed as
it is not empty.
So let's see if that helps.