Clarify removal of `puts` in CHANGELOG
The original language of the changed item was rather confusing; it almost made it sound like the deprecation of `puts` was being reversed, and that `puts` is now preferred over `log`. But it also said that both `puts` _and_ `log` were deprecated in 4.0. I don't think that's the case or the intention of this CHANGELOG point.

My edits aim to make clear that `puts` has been deprecated, and the desire of the Cucumber Team is that `log` should be used in step definitions instead.