Reactivate the Cucumber Compatibility Kit (#1579)
* Reactivate the CCK

* Make sure to use cucumber-create-meta 6.0.2

* Fix rubocop offense

* Disable fail-fast in github test workflow

* Remove empty tag_expressions from hook messages

The default for Cuccumber::Messages::Hook#tag_expression is nil. If
setting it to an empty string, it will be serialized as an empty string
in the messages. That is responsible for some example in the CCK to
fail.

By setting it to nil rather than an empty stirng, it is not serialized
anymore, thus it matches the examples of the compatibility kit.

* Add an entry in the CHANGELOG.md