Remove the use of the 'genki-ruby-terminfo' gem
The gem is not a dependency of cucumber, but its installation was
automatically detected by code.

This gem is not very popular, and unmaintained, and is used only to make
sure 256 colors are available on the user's terminal.

When the gem is not installed, it fallback to defining the 'grey' method
anyway.

Removing that gem result in declaring the 'grey' method as part of
'Cucumber::Term::ANSIColor': it did not make sense anymore to keep it
away from that module which already defined all other term colors.